Report DESCRIPTION-level issues.

Report issues with DESCRIPTION-level metadata of packages.

Usage

issues_descriptions(meta = meta_packages())

Arguments

meta

A data frame with R-universe package check results returned by meta_checks().

Value

A named list of information about packages which do not comply with DESCRPTION checks. Each name is a package name, and each element contains specific information about non-compliance.

Details

issues_descriptions() reports specific issues in the DESCRIPTION-level metadata of packages:

Security advisories at https://github.com/RConsortium/r-advisory-database.

Licenses that cannot be verified as free and open-source.

The presence of a "Remotes" field.

Package issues

Functions like issues_versions() and issues_descriptions() perform health checks for all packages in R-multiverse. For a complete list of checks, see the issues_*() functions listed at https://r-multiverse.org/multiverse.internals/reference/index.html. record_versions() updates the version number history of releases in R-multiverse, and record_issues() gathers together all the issues about R-multiverse packages.

See also

Examples

  meta <- meta_packages(repo = "https://wlandau.r-universe.dev")
  issues <- issues_descriptions(meta = meta)
  str(issues)
#>  list()