Report package check synchronization issues.

Ensure the reported R-universe checks are synchronized. Report the packages whose checks have not been synchronized.

Usage

issues_synchronization(meta = meta_packages(), verbose = FALSE)

Arguments

meta

Package metadata from meta_packages().

verbose

TRUE to print progress messages, FALSE otherwise.

Value

A tibble with one row per package and the following columns:

package: Name of the package.

synchronization: Synchronization status: "success" if the checks are synchronized, "incomplete" if checks are still running on R-universe GitHub Actions, and "recent" if the package was last published so recently that downstream checks may not have started yet.

Details

R-universe automatically rechecks downstream packages if an upstream dependency increments its version number. R-multiverse needs to wait for these downstream checks to finish before it makes decisions about accepting packages into Production. issues_synchronization() scrapes the GitHub Actions API to find out if any R-universe checks are still running for a package. In addition, to give rechecks enough time to post on GitHub Actions, it flags packages published within the last 5 minutes.

See also

Examples

  if (FALSE) { # \dontrun{
  meta <- meta_packages(repo = "https://wlandau.r-universe.dev")
  issues_synchronization(meta)
  } # }