List package metadata

List package metadata in an R universe.

Usage

meta_packages(repo = "https://community.r-multiverse.org")

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

Value

A data frame with one row per package and columns with package metadata.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package     version
#> autometric     673686b9ca3f41f657d09b74     autometric  0.1.2.9000
#> crew           6737776408d805b65aa894e8           crew 0.10.2.9000
#> crew.aws.batch 6736876c08d805b65aa64edf crew.aws.batch       0.0.7
#> crew.cluster   6736875a08d805b65aa64e44   crew.cluster       0.3.3
#> mirai          6737412008d805b65aa84905          mirai  1.3.1.9000
#> nanonext       6736872508d805b65aa64c46       nanonext  1.3.2.9000
#>                           license                                remotesha
#> autometric     MIT + file LICENSE 13ec8d2cfba955081cf3e19eb5e74af07ba89bef
#> crew           MIT + file LICENSE 04beece4e8b12a24ff004e48c1398a9843324a64
#> crew.aws.batch MIT + file LICENSE fc331371bc8c3f7722229849d77be0a031f34bdd
#> crew.cluster   MIT + file LICENSE ffa70aa9f5f2b9ca512223c7bd66c71c5d3cef1b
#> mirai                  GPL (>= 3) a40319e456cdf1f685d40c1845071ba7dd8d4da3
#> nanonext               GPL (>= 3) d713c85d02f1b9fac92796de78256b4d6c571a8a
#>                needscompilation                           md5sum _type
#> autometric                  yes 0036353f407cdf59507766ff998d9583   src
#> crew                         no e16680264b828c5f6eed5135ade0b439   src
#> crew.aws.batch               no 7fb448c7d8c699642b05ec2f1f78cf40   src
#> crew.cluster                 no 46cec9579c6123ce6db24710d5b03030   src
#> mirai                        no d8c9a7007ef27d5eaf77a85a6873885f   src
#> nanonext                    yes b0db16880501a3eceedb9575947283e3   src
#>                                                                         _fileid
#> autometric     8d19076199e56fa9e9da1a466c48d76ad9fc9d1f02d654b13e98603baf57a9f5
#> crew           01889039297c68968e93ffd382370bd1aa38a00411302d2b71389eecc2889187
#> crew.aws.batch 64d9500678267da53f5cb73afec73f682a6750dc8989893e8bc64a8a95e0f144
#> crew.cluster   b656b6cd4e121b8290811e63c253f7b5120c6793ff101501fce4ae43b1fea15e
#> mirai          09b2a18e34d673fdff1ab9ea5abe58b3fa6b62d3eee808c4e8c8e54e4574325c
#> nanonext       dd17b873a9f45365acf94e19c428e0ed0fb7c2cbf1c7fa0ff566b745668df982
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      _dependencies
#> autometric                                                                                                                                                                                                                                                                                                                                                                                                      R, graphics, utils, grDevices, ps, tinytest, >= 3.5.0, NA, NA, NA, NA, NA, Depends, Imports, Imports, Suggests, Suggests, Suggests
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, autometric, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 1.3.0, >= 1.3.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 0.1.0, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                                                      R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.10.2, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                                                           R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.10.2, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                                                                   R, nanonext, cli, litedown, parallel, promises, >= 3.6, >= 1.3.0, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                                                                   R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#>                filesize distro
#> autometric       273451  noble
#> crew            1063951  noble
#> crew.aws.batch   564828  noble
#> crew.cluster     390071  noble
#> mirai            409284  noble
#> nanonext        1078080  noble
#>                                                                          sha256
#> autometric     8d19076199e56fa9e9da1a466c48d76ad9fc9d1f02d654b13e98603baf57a9f5
#> crew           01889039297c68968e93ffd382370bd1aa38a00411302d2b71389eecc2889187
#> crew.aws.batch 64d9500678267da53f5cb73afec73f682a6750dc8989893e8bc64a8a95e0f144
#> crew.cluster   b656b6cd4e121b8290811e63c253f7b5120c6793ff101501fce4ae43b1fea15e
#> mirai          09b2a18e34d673fdff1ab9ea5abe58b3fa6b62d3eee808c4e8c8e54e4574325c
#> nanonext       dd17b873a9f45365acf94e19c428e0ed0fb7c2cbf1c7fa0ff566b745668df982
#>                     remotes
#> autometric             NULL
#> crew                   NULL
#> crew.aws.batch wlandau/crew
#> crew.cluster   wlandau/crew
#> mirai                  NULL
#> nanonext               NULL