Record nonstandard licenses

R-multiverse packages must have valid free and open-source (FOSS) licenses to protect the intellectual property rights of the package owners (c.f. https://en.wikipedia.org/wiki/Free_and_open-source_software). record_nonstandard_licenses() records packages with nonstandard licenses.

Usage

record_nonstandard_licenses(
  repo = "https://community.r-multiverse.org",
  path = "nonstandard_licenses.json"
)

Arguments

repo

Character string, URL of the repository.

path

Character string, output path to write JSON data with the names and licenses of packages with non-standard licenses.

Value

NULL (invisibly). Called for its side effects.