Review an R-multiverse contribution

Review a pull request to add packages to R-multiverse.

Usage

review_pull_request(
  owner = "r-multiverse",
  repo = "contributions",
  number,
  advisories = character(0L),
  organizations = character(0L)
)

Arguments

owner

Character of length 1, name of the package repository owner.

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

number

Positive integer of length 1, index of the pull request in the repo.

advisories

Character vector of names of packages with advisories in the R Consortium Advisory Database.

organizations

Character vector of names of GitHub organizations. Pull requests from authors who are not members of at least one of these organizations will be flagged for manual review.

Value

NULL (invisibly).

See also