Update production

Update the production universe.

Usage

update_production(
  path_production,
  path_community,
  repo_production = "https://production.r-multiverse.org",
  repo_community = "https://community.r-multiverse.org",
  days_notice = 28L,
  mock = NULL
)

Arguments

path_production

Character string, directory path to the source files of the production universe.

path_community

Character string, directory path to the source files of the community universe.

repo_production

Character string, URL of the production universe.

repo_community

Character string, URL of the community universe.

days_notice

Integer scalar, number of days between the detection of a production issue and removal from the production universe.

mock

For testing purposes only, a named list of data frames for inputs to various intermediate functions.

Value

NULL (invisibly)

Details

update_production() controls how packages enter and leave the production universe. It updates the production packages.json manifest depending on the contents of the community universe and issues with package checks. There are 3 phases:

Demote packages: packages with any check issues in production are given days_notice days to fix the problems. If the problems are fixed on time, then the package stays in production, and the notice period resets (the next problem is given the full days_notice notice period from scratch). Otherwise, if there are still issues after days_notice days, then the package is removed from the packages.json manifest and added to a special removing.json manifest in production. removing.json ensures that the builds are actually removed from production before the package can be promoted again. That way, if the next automatic promotion fails production checks, no build will be available to install from https://production.r-multiverse.org.

Clear removals: a demoted package stays in removing.json until the builds are removed from https://production.r-multiverse.org. After the builds are gone, the package is removed from removing.json so it can be promoted again. No build will become available again until the package passes all production checks.

Promote packages: a package in the community universe is moved to production if:

It passes description checks from issues_descriptions().

It passes version checks from issues_versions().

The package is available to install from https://community.r-multiverse.org.

The package is not in removing.json. To promote the package, an entry is created in the production packages.json with the Git hash of the latest release.

Examples

if (FALSE) { # \dontrun{
url_production = "https://github.com/r-multiverse/production"
url_community = "https://github.com/r-multiverse/community"
path_production <- tempfile()
path_community <- tempfile()
gert::git_clone(url = url_production, path = path_production)
gert::git_clone(url = url_community, path = path_community)
update_production(
  repo_production = "https://production.r-multiverse.org",
  repo_community = "https://community.r-multiverse.org",
  path_production = path_production,
  path_community = path_community,
  days_notice = 28L
)
} # }