Update the package status repository

Update the repository which reports the status on individual packages.

Usage

update_status(
  path_status,
  path_staging,
  path_community,
  repo_staging = "https://staging.r-multiverse.org",
  repo_community = "https://community.r-multiverse.org",
  mock = NULL
)

Arguments

path_status

Character string, directory path to the source files of the package status repository.

path_staging

Character string, directory path to the source files of the staging universe.

path_community

Character string, directory path to the source files of the community universe.

repo_staging

Character string, URL of the staging universe.

repo_community

Character string, URL of the community universe.

mock

For testing purposes only, a named list of data frames for inputs to various intermediate functions.

See also

Examples

if (FALSE) { # \dontrun{
url_staging = "https://github.com/r-multiverse/staging"
url_community = "https://github.com/r-multiverse/community"
path_status <- tempfile()
path_staging <- tempfile()
path_community <- tempfile()
gert::git_clone(url = url_staging, path = path_staging)
gert::git_clone(url = url_community, path = path_community)
update_status(
  path_status = path_status,
  path_staging = path_staging,
  path_community = path_community,
  repo_staging = "https://staging.r-multiverse.org",
  repo_community = "https://community.r-multiverse.org"
)
writeLines(
  readLines(
    file.path(path_status, "community", "multiverse.internals.html")
  )
)
writeLines(
  readLines(
    file.path(path_status, "community", "multiverse.internals.xml")
  )
)
} # }