List package metadata

List package metadata in an R universe.

Usage

meta_packages(repo = "https://community.r-multiverse.org")

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

Value

A data frame with one row per package and columns with package metadata.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package    version
#> crew           66f1888d363ce6fc663b8425           crew 0.9.5.9011
#> crew.aws.batch 66df04a3aa1e0109854d42c3 crew.aws.batch 0.0.6.9005
#> crew.cluster   66df04baaa1e0109854d432e   crew.cluster 0.3.2.9004
#> mirai          66ed8407363ce6fc663576cb          mirai 1.2.0.9020
#> nanonext       66ed6ad7363ce6fc66351b79       nanonext 1.2.1.9020
#>                           license                                remotesha
#> crew           MIT + file LICENSE 378e3c103d959f71ff6cd7c4672171ba5f468b73
#> crew.aws.batch MIT + file LICENSE 9d71570b75de8679176480227e976c3ce7ba4eac
#> crew.cluster   MIT + file LICENSE f916bfb3b650a2fcf88bce1ac97e8cece3ce36cc
#> mirai                  GPL (>= 3) 12479a508ac237e89e2e471caa931e137810f38e
#> nanonext               GPL (>= 3) 790f8bab32e369c4494e247b7790ed2055da57f5
#>                needscompilation                           md5sum _type
#> crew                         no 956faa7f7bfe5167e5135c15be07b6f7   src
#> crew.aws.batch               no bb1a492f402ac870cf0276af085893d1   src
#> crew.cluster                 no 9e5a395703c6928caf6ace682b5d2c75   src
#> mirai                        no 10fae312edc9ed1d69bb841aaf3fca2a   src
#> nanonext                    yes 22ff6de4a3e01477b5897f8a33aedb26   src
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      _dependencies
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 1.1.0, >= 1.1.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                  R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.9.5.9007, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                       R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.9.5.9007, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                              R, nanonext, cli, litedown, parallel, promises, >= 3.6, >= 1.2.1.9020, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                                   R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#>                filesize distro              remotes
#> crew            1015260  noble                 NULL
#> crew.aws.batch   564779  noble                 NULL
#> crew.cluster     397391  noble                 NULL
#> mirai            397378  noble shikokuchuo/nanonext
#> nanonext        1070660  noble                 NULL