List package metadata

List package metadata in an R universe.

Usage

meta_packages(repo = "https://multiverse.r-multiverse.org")

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://multiverse.r-multiverse.org".

Value

A data frame with one row per package and columns with package metadata.

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                        _id        package    version            license
#> 1 666319a14e86770016661a28           crew 0.9.3.9002 MIT + file LICENSE
#> 2 66580c6dd014ca0014e5afc2 crew.aws.batch 0.0.5.9000 MIT + file LICENSE
#> 3 665179088d83a20014a4037a   crew.cluster      0.3.1 MIT + file LICENSE
#> 4 6666d6575e691000165322b1          mirai 1.1.0.9000         GPL (>= 3)
#> 5 6666d6865e691000165322fe       nanonext 1.1.0.9000         GPL (>= 3)
#>                                  remotesha needscompilation
#> 1 eafad0276c06dec2344da2f03596178c754c8b5e               no
#> 2 4d9e5b44e2942d119af963339c48d134e84de458               no
#> 3 d4ac61fd9a1d9539088ffebdadcd4bb713c25ee1               no
#> 4 7015695b7ef82f82ab3225ac2d226b2c8f298097               no
#> 5 85dd672a44a92c890eb40ea9ebab7a4e95335c2f              yes
#>                             md5sum _type
#> 1 642e9ac93d39d462ee6f95d8522afbe1   src
#> 2 23c2733447eae95614fd6af1889b57e8   src
#> 3 5707d786c1aa3848fd37e6bf4598ea4c   src
#> 4 3c92053c75031ec0b976b0a15185e3a0   src
#> 5 1507b3a27da7dff5d9acbf8ef181ad78   src
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           _dependencies
#> 1 R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 0.12.0, >= 0.12.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 2                                                                                                                                         R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.8.0, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 3                                                                                                                                                                            R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.8.0, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 4                                                                                                                                                                                                                                                                                                                                                     R, nanonext, knitr, markdown, parallel, promises, >= 3.6, >= 1.1.0.9000, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> 5                                                                                                                                                                                                                                                                                                                                                                                             R, later, later, knitr, markdown, >= 3.5, NA, NA, NA, NA, Depends, LinkingTo, Imports, Suggests, Suggests
#>   distro
#> 1  noble
#> 2  noble
#> 3  noble
#> 4  noble
#> 5  noble