List package metadata

List package metadata in an R universe.

Usage

meta_packages(repo = "https://community.r-multiverse.org")

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

Value

A data frame with one row per package and columns with package metadata.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package    version
#> crew           66b0d2a9b506c5b6321da186           crew 0.9.5.9007
#> crew.aws.batch 66ad95df76d89de6b1761df2 crew.aws.batch 0.0.6.9005
#> crew.cluster   66ad95a676d89de6b1761db0   crew.cluster 0.3.2.9004
#> mirai          66c4eb95b9e16199e5b3b184          mirai 1.2.0.9001
#> nanonext       66c4faabb9e16199e5b3bc9f       nanonext 1.2.1.9001
#>                           license                                remotesha
#> crew           MIT + file LICENSE 96a190cef8712ad39561d591afa0d52a4708a53d
#> crew.aws.batch MIT + file LICENSE ca493bc9c61e4886a64f9b4da5778e7603fde3d1
#> crew.cluster   MIT + file LICENSE a35c967004cccd947e7011fdf6f7fef674154c48
#> mirai                  GPL (>= 3) 29c0bfed81e4249de4ed79e8db192e6e309b193d
#> nanonext               GPL (>= 3) f4446fe2f967d84fd2bd8d47f2aaaea7e3b1478a
#>                needscompilation                           md5sum _type
#> crew                         no 70e87883228b7408bb7f083949c32158   src
#> crew.aws.batch               no 24edea5adb6ffb53f0aa545cc2ccf7d9   src
#> crew.cluster                 no 2195743ae468ef16038889c4f6bc7cfe   src
#> mirai                        no 3baa17a74a8a42eecb2bebc39fd3814a   src
#> nanonext                    yes c6bd4e44f0af5a9d99915b5685a6d209   src
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      _dependencies
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 1.1.0, >= 1.1.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                  R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.9.5.9007, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                       R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.9.5.9007, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                                                      R, nanonext, litedown, parallel, promises, >= 3.6, >= 1.2.0, NA, NA, NA, Depends, Imports, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                                   R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#>                filesize distro
#> crew            1015991  noble
#> crew.aws.batch   560677  noble
#> crew.cluster     395481  noble
#> mirai            387490  noble
#> nanonext        1065595  noble