List package metadata

List package metadata in an R universe.

Usage

meta_packages(
  repo = "https://community.r-multiverse.org",
  fields = c("Version", "License", "Remotes", "RemoteSha")
)

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

fields

Character string of fields to query.

Value

A data frame with one row per package and columns with package metadata.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package     version
#> autometric     675d0512a59ade8d046e75f7     autometric  0.1.2.9000
#> crew           6762ec1f89f7bbc5d2ac0911           crew 0.10.2.9001
#> crew.aws.batch 6762412989f7bbc5d2a9eb0e crew.aws.batch  0.0.7.9000
#> crew.cluster   676240d289f7bbc5d2a9e9dd   crew.cluster  0.3.3.9000
#> mirai          6751806f2b9dbe1a26600cc7          mirai  1.3.1.9000
#> nanonext       674ec20e31a30401b85331d0       nanonext  1.4.0.9000
#>                           license                                remotesha
#> autometric     MIT + file LICENSE 13ec8d2cfba955081cf3e19eb5e74af07ba89bef
#> crew           MIT + file LICENSE 367222969e86ebb866d81a058d1b0fd7003178c6
#> crew.aws.batch MIT + file LICENSE caa4752f528f15510eccf5fab0cecb1cdfb5fa4b
#> crew.cluster   MIT + file LICENSE 02e3d32812394fb36bb6d7c76146af122fa4162c
#> mirai                  GPL (>= 3) 2020f0a5dfed1ad48a3afd5de0668d929242f839
#> nanonext               GPL (>= 3) 4814abd0d47977ead3773da44e1e00f4a6c45e0a
#>                needscompilation                           md5sum _type
#> autometric                  yes 9c6c39b65c4e0050cad36dda0e8ccfe6   src
#> crew                         no 1ca625bce5776d03f23775832a97135f   src
#> crew.aws.batch               no 4e27f1c6d9f12797aa664c0e67134465   src
#> crew.cluster                 no 4503771cac9d9c00f65e4658554d59da   src
#> mirai                        no ff7c9574bbdcb9807f46561cdd77a734   src
#> nanonext                    yes d479fa3d4bf045bfea098af73861f097   src
#>                                                                         _fileid
#> autometric     a847b302b35d5df48486e0c8362fa21db7d1efde2c11a212876c745e9bd80b10
#> crew           8c346d119ec2e731ce3a1a9cba80c4752a367117cb80787892d72f47d0d50a10
#> crew.aws.batch f2ee88676843c36e422748a9d898ec24a0dd4485b2f5c4d7285ebf9cb0309e4b
#> crew.cluster   bd40a9b3d82cba25e11b9e7736e5d79de4b5d8649f32b87f5f0e0aaa54170d74
#> mirai          ad690ed7c1601446f2e742a49cbd0f894de5a23e1a778b8ed728c516d898276a
#> nanonext       b901ac0419c88f716c0d0f0786910c723f4ac365c2c35f176d49317fec56fab9
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      _dependencies
#> autometric                                                                                                                                                                                                                                                                                                                                                                                                      R, graphics, utils, grDevices, ps, tinytest, >= 3.5.0, NA, NA, NA, NA, NA, Depends, Imports, Imports, Suggests, Suggests, Suggests
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, autometric, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 1.3.0, >= 1.3.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 0.1.0, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                                                      R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.10.2, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                                                           R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.10.2, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                                                                   R, nanonext, cli, litedown, parallel, promises, >= 3.6, >= 1.3.0, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                                                                   R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#>                filesize distro
#> autometric       273449  noble
#> crew            1063948  noble
#> crew.aws.batch   564858  noble
#> crew.cluster     390048  noble
#> mirai            410083  noble
#> nanonext        1079031  noble
#>                                                                          sha256
#> autometric     a847b302b35d5df48486e0c8362fa21db7d1efde2c11a212876c745e9bd80b10
#> crew           8c346d119ec2e731ce3a1a9cba80c4752a367117cb80787892d72f47d0d50a10
#> crew.aws.batch f2ee88676843c36e422748a9d898ec24a0dd4485b2f5c4d7285ebf9cb0309e4b
#> crew.cluster   bd40a9b3d82cba25e11b9e7736e5d79de4b5d8649f32b87f5f0e0aaa54170d74
#> mirai          ad690ed7c1601446f2e742a49cbd0f894de5a23e1a778b8ed728c516d898276a
#> nanonext       b901ac0419c88f716c0d0f0786910c723f4ac365c2c35f176d49317fec56fab9
#>                foss
#> autometric     TRUE
#> crew           TRUE
#> crew.aws.batch TRUE
#> crew.cluster   TRUE
#> mirai          TRUE
#> nanonext       TRUE