List package metadata

List package metadata in an R universe.

Usage

meta_packages(
  repo = "https://community.r-multiverse.org",
  fields = c("Version", "License", "Remotes", "RemoteSha")
)

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

fields

Character string of fields to query.

Value

A data frame with one row per package and columns with package metadata.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package    version
#> autometric     6784912a040e85d8a67224db     autometric 0.1.2.9000
#> crew           67a38e7683ace1b0aa3bf307           crew 1.0.0.9001
#> crew.aws.batch 67a39e3e83ace1b0aa3bfc75 crew.aws.batch      0.0.8
#> crew.cluster   67a39e1b83ace1b0aa3bfc3c   crew.cluster 0.3.4.9000
#> mirai          67a5df7c4083faa4562d5571          mirai 2.1.0.9000
#> nanonext       67a5612e4083faa4562bc609       nanonext 1.5.0.9000
#>                           license                                remotesha
#> autometric     MIT + file LICENSE 13ec8d2cfba955081cf3e19eb5e74af07ba89bef
#> crew           MIT + file LICENSE 9bc7c2c8b3cd60a7b64ff128225a2fb8ec360cb0
#> crew.aws.batch MIT + file LICENSE 136fedb9df30123b7c216ac75a66f29e8e94f2c1
#> crew.cluster   MIT + file LICENSE f9ecd38a1b8736157b06635f5e0cb43cb80c09a4
#> mirai                  GPL (>= 3) 67f95d0eee0b35637d02589fc4ecd78303ada586
#> nanonext               GPL (>= 3) c52954a35fa7fe10de340811a1405897e603ba9a
#>                needscompilation                           md5sum _type
#> autometric                  yes c5c4fb3ba112cc8217c6754ad8879508   src
#> crew                         no 6f186b98fafac1fb3d673c33f894714f   src
#> crew.aws.batch               no 41cb06e4ff71a43b97fcba6240ab1ec8   src
#> crew.cluster                 no 1ffb744c7d5f78b7d3d0a1ea059a3d20   src
#> mirai                        no 503c6cb586b5c8370f961ae0d89af473   src
#> nanonext                    yes 8fd414d456c93462b1c9711b754c2e3f   src
#>                                                                         _fileid
#> autometric     7323e5490f368203f3360e142c6b594b7bfa8a298a1e83490898bbfefc8dee7b
#> crew           049fafa4cd173b0fdfb0826593ae309115804a5062408f60cb8e79d9785b979f
#> crew.aws.batch 5653359fc41384cebab63acf150566affa1c81f257cd773ab6dc8ba47607bf94
#> crew.cluster   f56df6cc525926983aef4ddc8633fc0186fa790c93566c6c4e681fce623e6bce
#> mirai          6bdc4d7615e299a484b16b0a60d5b0c88520cb7baddb5cfc6e915755d8f306cf
#> nanonext       1661ec9da9d2fd90eea79b0df6a74b1f88d45208900d210871ac426650f669b3
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      _dependencies
#> autometric                                                                                                                                                                                                                                                                                                                                                                                                      R, graphics, utils, grDevices, ps, tinytest, >= 3.5.0, NA, NA, NA, NA, NA, Depends, Imports, Imports, Suggests, Suggests, Suggests
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, autometric, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 2.0.1, >= 1.4.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 0.1.0, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                                                       R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 1.0.0, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                                                            R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 1.0.0, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                                                                   R, nanonext, cli, litedown, parallel, promises, >= 3.6, >= 1.5.0, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                                                                   R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#>                filesize distro
#> autometric       273551  noble
#> crew            1083488  noble
#> crew.aws.batch   572616  noble
#> crew.cluster     389512  noble
#> mirai            395358  noble
#> nanonext        1039134  noble
#>                                                                          sha256
#> autometric     7323e5490f368203f3360e142c6b594b7bfa8a298a1e83490898bbfefc8dee7b
#> crew           049fafa4cd173b0fdfb0826593ae309115804a5062408f60cb8e79d9785b979f
#> crew.aws.batch 5653359fc41384cebab63acf150566affa1c81f257cd773ab6dc8ba47607bf94
#> crew.cluster   f56df6cc525926983aef4ddc8633fc0186fa790c93566c6c4e681fce623e6bce
#> mirai          6bdc4d7615e299a484b16b0a60d5b0c88520cb7baddb5cfc6e915755d8f306cf
#> nanonext       1661ec9da9d2fd90eea79b0df6a74b1f88d45208900d210871ac426650f669b3
#>                foss
#> autometric     TRUE
#> crew           TRUE
#> crew.aws.batch TRUE
#> crew.cluster   TRUE
#> mirai          TRUE
#> nanonext       TRUE