List package metadata

List package metadata in an R universe.

Usage

meta_packages(repo = "https://community.r-multiverse.org")

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

Value

A data frame with one row per package and columns with package metadata.

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package    version
#> crew           66a159e867947a312ef110b0           crew 0.9.5.9003
#> crew.aws.batch 66a1b8be67947a312ef1aef1 crew.aws.batch 0.0.6.9004
#> crew.cluster   66a1b8aa67947a312ef1ae62   crew.cluster 0.3.2.9003
#> mirai          66a233f7309255ca4e168f5e          mirai 1.1.1.9007
#> nanonext       66a17e0967947a312ef174ac       nanonext 1.1.1.9007
#>                           license                                remotesha
#> crew           MIT + file LICENSE e62928e1f10fea46b3c6c370afdba208af9bac49
#> crew.aws.batch MIT + file LICENSE 16ce8f9a43387b6b54e2918806d8e6121927bdf3
#> crew.cluster   MIT + file LICENSE 000ae593700e72369e216641e7ad83c180f45da2
#> mirai                  GPL (>= 3) 8a071189e21769472d3a3a355b8992d82308df56
#> nanonext               GPL (>= 3) 852c2a0848d6d1ec64f614bcea2879134e419e72
#>                needscompilation                           md5sum _type
#> crew                         no da48854ec2e9f6993d09230284f7dbe4   src
#> crew.aws.batch               no cb11b8f7dfc513ddfb3a9d1fe099c3e9   src
#> crew.cluster                 no d5825710039948c89160973b744c22b1   src
#> mirai                        no 9d83413c0ddb64e95b111cbfa46fd1eb   src
#> nanonext                    yes bdc92a068da54e0500d52ed29fa761eb   src
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        _dependencies
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 0.12.0, >= 0.12.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                    R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.9.5.9003, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                         R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.9.5.9003, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                              R, nanonext, knitr, markdown, parallel, promises, >= 3.6, >= 1.1.1.9007, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                R, knitr, later, markdown, promises, >= 3.6, NA, NA, NA, NA, Depends, Suggests, Suggests, Suggests, Enhances
#>                filesize distro
#> crew             992518  noble
#> crew.aws.batch   557895  noble
#> crew.cluster     386910  noble
#> mirai            390241  noble
#> nanonext        1070690  noble