List package metadata

List package metadata in an R universe.

Usage

meta_packages(
  repo = "https://community.r-multiverse.org",
  fields = c("Version", "License", "Remotes", "RemoteSha")
)

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

fields

Character string of fields to query.

Value

A data frame with one row per package and columns with package metadata.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package     version
#> autometric     6784912a040e85d8a67224db     autometric  0.1.2.9000
#> crew           6785160d040e85d8a6735c43           crew 0.10.2.9001
#> crew.aws.batch 6762412989f7bbc5d2a9eb0e crew.aws.batch  0.0.7.9000
#> crew.cluster   676240d289f7bbc5d2a9e9dd   crew.cluster  0.3.3.9000
#> mirai          6786ad88040e85d8a6762fb0          mirai  2.0.0.9000
#> nanonext       6786f225040e85d8a6768e65       nanonext  1.4.0.9002
#>                           license                                remotesha
#> autometric     MIT + file LICENSE 13ec8d2cfba955081cf3e19eb5e74af07ba89bef
#> crew           MIT + file LICENSE b6207f2b80b12c24c89e0b4459b87eada9a2d3f7
#> crew.aws.batch MIT + file LICENSE caa4752f528f15510eccf5fab0cecb1cdfb5fa4b
#> crew.cluster   MIT + file LICENSE 02e3d32812394fb36bb6d7c76146af122fa4162c
#> mirai                  GPL (>= 3) feac10b651c20df586a9a90f3667c4825d0a1f45
#> nanonext               GPL (>= 3) 00c4d35f7de1795b65a7b564e9e34c3c593cd7bf
#>                needscompilation                           md5sum _type
#> autometric                  yes c5c4fb3ba112cc8217c6754ad8879508   src
#> crew                         no 27acf03720b21c778f804df9778012a3   src
#> crew.aws.batch               no 4e27f1c6d9f12797aa664c0e67134465   src
#> crew.cluster                 no 4503771cac9d9c00f65e4658554d59da   src
#> mirai                        no 9194ddbf5e319a545269bb8808c82fca   src
#> nanonext                    yes 765dfe93cecb5a00a4c9158556c3f24d   src
#>                                                                         _fileid
#> autometric     7323e5490f368203f3360e142c6b594b7bfa8a298a1e83490898bbfefc8dee7b
#> crew           eb4743e20f3ed697d1df9012b66ed1f89510f3b1ead5a9d1d99f4e2dd020d251
#> crew.aws.batch f2ee88676843c36e422748a9d898ec24a0dd4485b2f5c4d7285ebf9cb0309e4b
#> crew.cluster   bd40a9b3d82cba25e11b9e7736e5d79de4b5d8649f32b87f5f0e0aaa54170d74
#> mirai          b369fd80e28747375135db574447706ad05e1775f9faa383c6ba5d9018646b42
#> nanonext       f8ed9d36b78ca142acf25a42bd99f0572fbded02b8bc09b2cd81d0e8b927f448
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      _dependencies
#> autometric                                                                                                                                                                                                                                                                                                                                                                                                      R, graphics, utils, grDevices, ps, tinytest, >= 3.5.0, NA, NA, NA, NA, NA, Depends, Imports, Imports, Suggests, Suggests, Suggests
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, autometric, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 1.3.0, >= 1.3.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 0.1.0, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                                                      R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.10.2, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                                                           R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.10.2, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                                                                   R, nanonext, cli, litedown, parallel, promises, >= 3.6, >= 1.4.0, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                                                                   R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#>                filesize distro
#> autometric       273551  noble
#> crew            1063741  noble
#> crew.aws.batch   564858  noble
#> crew.cluster     390048  noble
#> mirai            393214  noble
#> nanonext        1074685  noble
#>                                                                          sha256
#> autometric     7323e5490f368203f3360e142c6b594b7bfa8a298a1e83490898bbfefc8dee7b
#> crew           eb4743e20f3ed697d1df9012b66ed1f89510f3b1ead5a9d1d99f4e2dd020d251
#> crew.aws.batch f2ee88676843c36e422748a9d898ec24a0dd4485b2f5c4d7285ebf9cb0309e4b
#> crew.cluster   bd40a9b3d82cba25e11b9e7736e5d79de4b5d8649f32b87f5f0e0aaa54170d74
#> mirai          b369fd80e28747375135db574447706ad05e1775f9faa383c6ba5d9018646b42
#> nanonext       f8ed9d36b78ca142acf25a42bd99f0572fbded02b8bc09b2cd81d0e8b927f448
#>                foss
#> autometric     TRUE
#> crew           TRUE
#> crew.aws.batch TRUE
#> crew.cluster   TRUE
#> mirai          TRUE
#> nanonext       TRUE