List package metadata

List package metadata in an R universe.

Usage

meta_packages(repo = "https://community.r-multiverse.org")

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

Value

A data frame with one row per package and columns with package metadata.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package     version
#> autometric     673686b9ca3f41f657d09b74     autometric  0.1.2.9000
#> crew           673f993cd7b8e0821a82853e           crew 0.10.2.9001
#> crew.aws.batch 673ab78a15eefd6298975cd1 crew.aws.batch  0.0.7.9000
#> crew.cluster   673ab78d15eefd6298975cf6   crew.cluster  0.3.3.9000
#> mirai          674d98717139319cf9db09b6          mirai  1.3.1.9000
#> nanonext       674e5e9731a30401b851e60e       nanonext       1.4.0
#>                           license                                remotesha
#> autometric     MIT + file LICENSE 13ec8d2cfba955081cf3e19eb5e74af07ba89bef
#> crew           MIT + file LICENSE 1e5a11392a12fb044c90e2889fd641f382a92c5e
#> crew.aws.batch MIT + file LICENSE caa4752f528f15510eccf5fab0cecb1cdfb5fa4b
#> crew.cluster   MIT + file LICENSE 02e3d32812394fb36bb6d7c76146af122fa4162c
#> mirai                  GPL (>= 3) 5189d4fb748c3a2199aeb8561cf5f16c38b951b2
#> nanonext               GPL (>= 3) a7394f4505097a4e1a2078ac6bc05fa23b1ea7bf
#>                needscompilation                           md5sum _type
#> autometric                  yes 0036353f407cdf59507766ff998d9583   src
#> crew                         no 45cf55bb34c4de29d1d8326b63ae2c60   src
#> crew.aws.batch               no 4f37ef14acd9cf99669f253b71c44c7e   src
#> crew.cluster                 no a86f08e7b4878139937db391c0d30261   src
#> mirai                        no 9b2bbe763501cf0d9e3226d91b5192fa   src
#> nanonext                    yes 1ef9571568f40ca790530f138eda0564   src
#>                                                                         _fileid
#> autometric     8d19076199e56fa9e9da1a466c48d76ad9fc9d1f02d654b13e98603baf57a9f5
#> crew           f07dfc4346bdd8c53d1b9eb50ba7850949a2b35c811db1cb12a89601dcc87e50
#> crew.aws.batch 4136ab3659a29f0f302fd9ada5040039961faa48a9ed815bb2277ef678184585
#> crew.cluster   670ef139cc062ac15ef07498a737e0b0689602a96b13b31b21205136a5e829bf
#> mirai          16b35c70ed573a4a8812612ee815b5a16a65ff82df4fa86034eb7c9bc28a1577
#> nanonext       3e81fa5dfe3f54bc2590cb411f9fb67f5271a7818619a1a6804aa3d918c9dd20
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                      _dependencies
#> autometric                                                                                                                                                                                                                                                                                                                                                                                                      R, graphics, utils, grDevices, ps, tinytest, >= 3.5.0, NA, NA, NA, NA, NA, Depends, Imports, Imports, Suggests, Suggests, Suggests
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, autometric, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 1.3.0, >= 1.3.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 0.1.0, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                                                      R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.10.2, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                                                           R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.10.2, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                                                                   R, nanonext, cli, litedown, parallel, promises, >= 3.6, >= 1.3.0, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                                                                   R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#>                filesize distro
#> autometric       273451  noble
#> crew            1063985  noble
#> crew.aws.batch   564883  noble
#> crew.cluster     390068  noble
#> mirai            410076  noble
#> nanonext        1079023  noble
#>                                                                          sha256
#> autometric     8d19076199e56fa9e9da1a466c48d76ad9fc9d1f02d654b13e98603baf57a9f5
#> crew           f07dfc4346bdd8c53d1b9eb50ba7850949a2b35c811db1cb12a89601dcc87e50
#> crew.aws.batch 4136ab3659a29f0f302fd9ada5040039961faa48a9ed815bb2277ef678184585
#> crew.cluster   670ef139cc062ac15ef07498a737e0b0689602a96b13b31b21205136a5e829bf
#> mirai          16b35c70ed573a4a8812612ee815b5a16a65ff82df4fa86034eb7c9bc28a1577
#> nanonext       3e81fa5dfe3f54bc2590cb411f9fb67f5271a7818619a1a6804aa3d918c9dd20
#>                foss
#> autometric     TRUE
#> crew           TRUE
#> crew.aws.batch TRUE
#> crew.cluster   TRUE
#> mirai          TRUE
#> nanonext       TRUE