List package metadata

List package metadata in an R universe.

Usage

meta_packages(repo = "https://community.r-multiverse.org")

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

Value

A data frame with one row per package and columns with package metadata.

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                                     _id        package    version
#> crew           667dcc4cf124990014cc4564           crew 0.9.5.9000
#> crew.aws.batch 668eb7d11611f70012980cfe crew.aws.batch 0.0.6.9000
#> crew.cluster   668eb7bb1611f70012980cde   crew.cluster 0.3.2.9000
#> mirai          669a5f27e4dd7a0012484b5d          mirai 1.1.1.9002
#> nanonext       669a4ff4e4dd7a0012481e4b       nanonext 1.1.1.9003
#>                           license                                remotesha
#> crew           MIT + file LICENSE 4acbfe34ca51322d1418c42d38f9db88bcee0c84
#> crew.aws.batch MIT + file LICENSE 300819f6bf0e9186f5e9d5ae2166e3f322eeaf9c
#> crew.cluster   MIT + file LICENSE 32e8f913a5cad1bbb64ad7164973b11139e6c459
#> mirai                  GPL (>= 3) eaed1f2f6a78cc195103927a9623fd660e7ef060
#> nanonext               GPL (>= 3) ea41edf908035406ff130227e8bcb8681c286515
#>                needscompilation                           md5sum _type
#> crew                         no 2ef0420d4cc0e41f4e4d614cda7744fe   src
#> crew.aws.batch               no ff1a505b9c5933623af8910798ad25e7   src
#> crew.cluster                 no fab6d8cd5b2ea668a5cfd6f7409e0996   src
#> mirai                        no 4fd17f30d113f3555a1095b4d6135050   src
#> nanonext                    yes 0bda218191de269b336bac36305f75dc   src
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        _dependencies
#> crew           R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 0.12.0, >= 0.12.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.aws.batch                                                                                                                                         R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.9.5, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> crew.cluster                                                                                                                                                                              R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.9.5, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> mirai                                                                                                                                                                                                                                                                                                                                                              R, nanonext, knitr, markdown, parallel, promises, >= 3.6, >= 1.1.1.9003, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> nanonext                                                                                                                                                                                                                                                                                                                                                                                                R, knitr, later, markdown, promises, >= 3.5, NA, NA, NA, NA, Depends, Suggests, Suggests, Suggests, Enhances
#>                filesize distro
#> crew             990557  noble
#> crew.aws.batch   559151  noble
#> crew.cluster     383165  noble
#> mirai            387373  noble
#> nanonext        1062416  noble