List package metadata

List package metadata in an R universe.

Usage

meta_packages(
  repo = "https://community.r-multiverse.org",
  fields = c("Version", "License", "Remotes", "RemoteSha")
)

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

fields

Character string of fields to query.

Value

A data frame with one row per package and columns with package metadata. The most important columns are:

package: character vector of package names.

version: character vector of package versions in the repo.

license: character vector of license names.

remotesha: character vector of GitHub/GitLab commit hashes.

remotes: list of character vectors with dependencies in the Remotes: field of the DESCRIPTION files.

foss: TRUE if the package has a valid free open-source license, FALSE otherwise.

cran: character vector of versions. Each version is the version of the package that was on CRAN during the first day of the most recent R-multiverse staging period.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>          package                      _id    version            license
#> 1     autometric 67ac192d4083faa45638ecac 0.1.2.9000 MIT + file LICENSE
#> 2           crew 67a641b44083faa4562ddbbd 1.0.0.9001 MIT + file LICENSE
#> 3 crew.aws.batch 67ca9744d1c8653be19e7c9f      0.0.8 MIT + file LICENSE
#> 4   crew.cluster 67ca931cd1c8653be19e607f 0.3.4.9000 MIT + file LICENSE
#> 5          mirai 67cb3b21d1c8653be19f5879 2.1.0.9001         GPL (>= 3)
#> 6       nanonext 67cb3bb0d1c8653be19f58d2 1.5.1.9001         GPL (>= 3)
#> 7        proffer 67ca603ad1c8653be19d5f6b 0.2.2.9000 MIT + file LICENSE
#> 8          withr 67ca2f31d1c8653be19d2100 3.0.2.9001 MIT + file LICENSE
#>                                  remotesha needscompilation
#> 1 13ec8d2cfba955081cf3e19eb5e74af07ba89bef              yes
#> 2 7d460e26f4c45233bee15ff4db9564503ec4b21a               no
#> 3 136fedb9df30123b7c216ac75a66f29e8e94f2c1               no
#> 4 f9ecd38a1b8736157b06635f5e0cb43cb80c09a4               no
#> 5 8a086779fc111826c6f87f62110319b1ddad5962               no
#> 6 e08cf94d86b23eed4a9968eee0042f4bcdd9cb8f              yes
#> 7 161b9e8f8e528384deffa6b8dd6e7abeeb130531               no
#> 8 59da5977ea1688e1c3064ab0c3eddb84c0be17d7               no
#>                             md5sum _type
#> 1 af8b02ad07b6ca96b952ccd662061b87   src
#> 2 c8a78b74aa4c15fd96a9fb8259e6f78e   src
#> 3 6e791b49bfe825d5472297326e42762a   src
#> 4 4f0769a5fffd5c239f57dfc85b622e7a   src
#> 5 524d8e2e9a3815fbcaf1b9471168c05b   src
#> 6 60ac47e6ea3364fd298dcfddcf028c58   src
#> 7 522382709e8273f786cb314b0e3cb672   src
#> 8 619e359ed790074faaf7492ac48055f8   src
#>                                                            _fileid
#> 1 67c829a5a8577aae969e94da76b29bd44e40f0c73ea93eae824eb1cbfaf93973
#> 2 0af52280edb8f56cd4422e83045e0bbe94276e11fed4364a3a534319e45f4cd2
#> 3 8302d7f626617fde57c53fa0e2eafb295a2492279875e547f94a1c501b5c18e1
#> 4 48cb4b7c5616fb9072a479a9d8094f198b191aa1cff2a94564084128c53169b1
#> 5 4e258357ddc4f7e5e0c034fbf45e3564a602c8054233d456d47dad4e184aa6fc
#> 6 84a91b3916b1f47663cc212483bf1173592677c40f5f67e59665f8c07595f98c
#> 7 9678ad68f5064c050e65a4b5f647c0388b8f7ec7ac0e0785112dff67f7965855
#> 8 dc5a831bd8203dcb1466dd1574fd8589c32a801a829c55d3fb1f2dbac6588994
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         _dependencies
#> 1                                                                                                                                                                                                                                                                                                                                                                                                  R, graphics, utils, grDevices, ps, tinytest, >= 3.5.0, NA, NA, NA, NA, NA, Depends, Imports, Imports, Suggests, Suggests, Suggests
#> 2 R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, autometric, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 2.0.1, >= 1.4.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 0.1.0, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests, Suggests
#> 3                                                                                                                                                                       R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 1.0.0, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 4                                                                                                                                                                                                          R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 1.0.0, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 5                                                                                                                                                                                                                                                                                                                                                                                          R, nanonext, cli, litedown, parallel, promises, >= 3.6, >= 1.5.0, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                             R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#> 7                                                                                                                                                                                                                                 R, cli, parallelly, pingr, processx, profile, R.utils, RProtoBuf, utils, withr, testthat, >= 3.3.0, >= 2.0.0, >= 1.26.0, >= 2.0.1, >= 3.4.0, >= 1.0, NA, >= 0.4.14, NA, >= 3.0.2.9001, >= 2.1.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests
#> 8                                                                                                                                                                                                                                                                     R, graphics, grDevices, callr, DBI, knitr, methods, rlang, rmarkdown, RSQLite, testthat, >= 3.6.0, NA, NA, NA, NA, NA, NA, NA, >= 2.12, NA, >= 3.0.0, Depends, Imports, Imports, Suggests, Suggests, Suggests, Suggests, Suggests, Suggests, Suggests, Suggests
#>   filesize distro
#> 1   274721  noble
#> 2  1083488  noble
#> 3   571284  noble
#> 4   389621  noble
#> 5   400506  noble
#> 6  1029928  noble
#> 7   817696  noble
#> 8   284462  noble
#>                                                             sha256 foss  cran
#> 1 67c829a5a8577aae969e94da76b29bd44e40f0c73ea93eae824eb1cbfaf93973 TRUE 0.1.2
#> 2 0af52280edb8f56cd4422e83045e0bbe94276e11fed4364a3a534319e45f4cd2 TRUE 1.0.0
#> 3 8302d7f626617fde57c53fa0e2eafb295a2492279875e547f94a1c501b5c18e1 TRUE 0.0.8
#> 4 48cb4b7c5616fb9072a479a9d8094f198b191aa1cff2a94564084128c53169b1 TRUE 0.3.4
#> 5 4e258357ddc4f7e5e0c034fbf45e3564a602c8054233d456d47dad4e184aa6fc TRUE 2.1.0
#> 6 84a91b3916b1f47663cc212483bf1173592677c40f5f67e59665f8c07595f98c TRUE 1.5.0
#> 7 9678ad68f5064c050e65a4b5f647c0388b8f7ec7ac0e0785112dff67f7965855 TRUE 0.2.2
#> 8 dc5a831bd8203dcb1466dd1574fd8589c32a801a829c55d3fb1f2dbac6588994 TRUE 3.0.2