List package metadata

List package metadata in an R universe.

Usage

meta_packages(repo = "https://community.r-multiverse.org")

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

Value

A data frame with one row per package and columns with package metadata.

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>                        _id        package    version            license
#> 1 66747689035c3700149b0f54           crew      0.9.4 MIT + file LICENSE
#> 2 66580c6dd014ca0014e5afc2 crew.aws.batch 0.0.5.9000 MIT + file LICENSE
#> 3 667465db035c3700149b0073   crew.cluster      0.3.1 MIT + file LICENSE
#> 4 6674bae332b53b0016d0d76b          mirai 1.1.0.9004         GPL (>= 3)
#> 5 6675759ef2c7cc00121b1740       nanonext 1.1.0.9004         GPL (>= 3)
#>                                  remotesha needscompilation
#> 1 c02449ff1e0441314576541a71688374a4d1ecaa               no
#> 2 4d9e5b44e2942d119af963339c48d134e84de458               no
#> 3 1adfcebbb02bd06e4687e11f1fedfa238e822433               no
#> 4 86b4b7c4319e66d56b9197f98f3fc1bfd13b18cc               no
#> 5 5f5aff199de48fb7d481b20f0d94433c91dbc183              yes
#>                             md5sum _type
#> 1 ba90fac40eca6e29b3ed4332bcd74bcf   src
#> 2 23c2733447eae95614fd6af1889b57e8   src
#> 3 49615ca9f7a7e5913a9d74a7df12842d   src
#> 4 0a6c3908efef61189b8091799d14a65b   src
#> 5 5350379026b9b402571d20f11f6311af   src
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           _dependencies
#> 1 R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 0.12.0, >= 0.12.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 2                                                                                                                                         R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 0.8.0, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 3                                                                                                                                                                            R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 0.8.0, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 4                                                                                                                                                                                                                                                                                                                                                     R, nanonext, knitr, markdown, parallel, promises, >= 3.6, >= 1.1.0.9003, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> 5                                                                                                                                                                                                                                                                                                                                                                                          R, knitr, later, markdown, promises, >= 3.5, NA, NA, NA, NA, Depends, Suggests, Suggests, Suggests, Enhances
#>   distro
#> 1  noble
#> 2  noble
#> 3  noble
#> 4  noble
#> 5  noble