List package metadata

List package metadata in an R universe.

Usage

meta_packages(
  repo = "https://community.r-multiverse.org",
  fields = c("Version", "License", "Remotes", "RemoteSha")
)

Arguments

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

fields

Character string of fields to query.

Value

A data frame with one row per package and columns with package metadata. The most important columns are:

package: character vector of package names.

version: character vector of package versions in the repo.

license: character vector of license names.

remotesha: character vector of GitHub/GitLab commit hashes.

remotes: list of character vectors with dependencies in the Remotes: field of the DESCRIPTION files.

foss: TRUE if the package has a valid free open-source license, FALSE otherwise.

cran: character vector of versions. Each version is the version of the package that was on CRAN during the first day of the most recent R-multiverse staging period.

See also

Examples

meta_packages(repo = "https://wlandau.r-universe.dev")
#>          package                      _id    version            license
#> 1     autometric 67ac192d4083faa45638ecac 0.1.2.9000 MIT + file LICENSE
#> 2           crew 67a641b44083faa4562ddbbd 1.0.0.9001 MIT + file LICENSE
#> 3 crew.aws.batch 67a39e3e83ace1b0aa3bfc75      0.0.8 MIT + file LICENSE
#> 4   crew.cluster 67a39e1b83ace1b0aa3bfc3c 0.3.4.9000 MIT + file LICENSE
#> 5          mirai 67a5df7c4083faa4562d5571 2.1.0.9000         GPL (>= 3)
#> 6       nanonext 67a5612e4083faa4562bc609 1.5.0.9000         GPL (>= 3)
#>                                  remotesha needscompilation
#> 1 13ec8d2cfba955081cf3e19eb5e74af07ba89bef              yes
#> 2 7d460e26f4c45233bee15ff4db9564503ec4b21a               no
#> 3 136fedb9df30123b7c216ac75a66f29e8e94f2c1               no
#> 4 f9ecd38a1b8736157b06635f5e0cb43cb80c09a4               no
#> 5 67f95d0eee0b35637d02589fc4ecd78303ada586               no
#> 6 c52954a35fa7fe10de340811a1405897e603ba9a              yes
#>                             md5sum _type
#> 1 af8b02ad07b6ca96b952ccd662061b87   src
#> 2 c8a78b74aa4c15fd96a9fb8259e6f78e   src
#> 3 41cb06e4ff71a43b97fcba6240ab1ec8   src
#> 4 1ffb744c7d5f78b7d3d0a1ea059a3d20   src
#> 5 503c6cb586b5c8370f961ae0d89af473   src
#> 6 8fd414d456c93462b1c9711b754c2e3f   src
#>                                                            _fileid
#> 1 67c829a5a8577aae969e94da76b29bd44e40f0c73ea93eae824eb1cbfaf93973
#> 2 0af52280edb8f56cd4422e83045e0bbe94276e11fed4364a3a534319e45f4cd2
#> 3 5653359fc41384cebab63acf150566affa1c81f257cd773ab6dc8ba47607bf94
#> 4 f56df6cc525926983aef4ddc8633fc0186fa790c93566c6c4e681fce623e6bce
#> 5 6bdc4d7615e299a484b16b0a60d5b0c88520cb7baddb5cfc6e915755d8f306cf
#> 6 1661ec9da9d2fd90eea79b0df6a74b1f88d45208900d210871ac426650f669b3
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         _dependencies
#> 1                                                                                                                                                                                                                                                                                                                                                                                                  R, graphics, utils, grDevices, ps, tinytest, >= 3.5.0, NA, NA, NA, NA, NA, Depends, Imports, Imports, Suggests, Suggests, Suggests
#> 2 R, cli, data.table, getip, later, mirai, nanonext, processx, promises, ps, R6, rlang, stats, tibble, tidyselect, tools, utils, autometric, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, NA, NA, NA, >= 2.0.1, >= 1.4.0, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, >= 0.1.0, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests, Suggests
#> 3                                                                                                                                                                       R, cli, crew, paws.common, paws.compute, paws.management, R6, rlang, tibble, utils, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 3.1.0, >= 1.0.0, >= 0.7.0, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 4                                                                                                                                                                                                          R, crew, ps, lifecycle, R6, rlang, utils, vctrs, xml2, yaml, knitr, markdown, rmarkdown, testthat, >= 4.0.0, >= 1.0.0, NA, NA, NA, NA, NA, NA, NA, NA, >= 1.30, >= 1.1, >= 2.4, >= 3.0.0, Depends, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Imports, Suggests, Suggests, Suggests, Suggests
#> 5                                                                                                                                                                                                                                                                                                                                                                                          R, nanonext, cli, litedown, parallel, promises, >= 3.6, >= 1.5.0, NA, NA, NA, NA, Depends, Imports, Suggests, Suggests, Enhances, Enhances
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                             R, later, litedown, promises, >= 3.6, NA, NA, NA, Depends, Suggests, Suggests, Enhances
#>   filesize distro
#> 1   274721  noble
#> 2  1083488  noble
#> 3   572616  noble
#> 4   389512  noble
#> 5   395358  noble
#> 6  1039134  noble
#>                                                             sha256 foss   cran
#> 1 67c829a5a8577aae969e94da76b29bd44e40f0c73ea93eae824eb1cbfaf93973 TRUE  0.1.2
#> 2 0af52280edb8f56cd4422e83045e0bbe94276e11fed4364a3a534319e45f4cd2 TRUE 0.10.2
#> 3 5653359fc41384cebab63acf150566affa1c81f257cd773ab6dc8ba47607bf94 TRUE  0.0.7
#> 4 f56df6cc525926983aef4ddc8633fc0186fa790c93566c6c4e681fce623e6bce TRUE  0.3.3
#> 5 6bdc4d7615e299a484b16b0a60d5b0c88520cb7baddb5cfc6e915755d8f306cf TRUE  2.0.0
#> 6 1661ec9da9d2fd90eea79b0df6a74b1f88d45208900d210871ac426650f669b3 TRUE  1.4.0