Record the manifest of package versions.

Record the manifest of versions of packages and their hashes.

Usage

record_versions(
  versions = "versions.json",
  repo = "https://community.r-multiverse.org",
  current = multiverse.internals::get_current_versions(repo = repo)
)

Arguments

versions

Character of length 1, file path to a JSON manifest tracking the history of released versions of packages.

repo

Character of length 1, URL of the package repository. R-multiverse uses "https://community.r-multiverse.org".

current

A data frame of current versions and hashes of packages in repo. This argument is exposed for testing only.

Value

NULL (invisibly). Writes a package version manifest and a manifest of version issues as JSON files.

Details

This function tracks a manifest containing the current version, the current hash, the highest version ever released, and the hash of the highest version ever released. issues_versions() uses this information to determine whether the package complies with best practices for version numbers.

Package issues

Functions like issues_versions() and issues_descriptions() perform health checks for all packages in R-multiverse. For a complete list of checks, see the issues_*() functions listed at https://r-multiverse.org/multiverse.internals/reference/index.html. record_versions() updates the version number history of releases in R-multiverse, and record_issues() gathers together all the issues about R-multiverse packages.

Examples

  # R-multiverse uses https://community.r-multiverse.org as the repo.
  repo <- "https://wlandau.r-universe.dev" # just for testing and examples
  output <- tempfile()
  versions <- tempfile()
  # First snapshot:
  record_versions(
    versions = versions,
    repo = repo
  )
  readLines(versions)
#>  [1] "["                                                                 
#>  [2] "  {"                                                               
#>  [3] "    \"package\": \"autometric\","                                  
#>  [4] "    \"version_current\": \"0.1.2.9000\","                          
#>  [5] "    \"hash_current\": \"13ec8d2cfba955081cf3e19eb5e74af07ba89bef\""
#>  [6] "  },"                                                              
#>  [7] "  {"                                                               
#>  [8] "    \"package\": \"crew\","                                        
#>  [9] "    \"version_current\": \"1.0.0.9001\","                          
#> [10] "    \"hash_current\": \"7d460e26f4c45233bee15ff4db9564503ec4b21a\""
#> [11] "  },"                                                              
#> [12] "  {"                                                               
#> [13] "    \"package\": \"crew.aws.batch\","                              
#> [14] "    \"version_current\": \"0.0.8\","                               
#> [15] "    \"hash_current\": \"136fedb9df30123b7c216ac75a66f29e8e94f2c1\""
#> [16] "  },"                                                              
#> [17] "  {"                                                               
#> [18] "    \"package\": \"crew.cluster\","                                
#> [19] "    \"version_current\": \"0.3.4.9000\","                          
#> [20] "    \"hash_current\": \"f9ecd38a1b8736157b06635f5e0cb43cb80c09a4\""
#> [21] "  },"                                                              
#> [22] "  {"                                                               
#> [23] "    \"package\": \"mirai\","                                       
#> [24] "    \"version_current\": \"2.1.0.9001\","                          
#> [25] "    \"hash_current\": \"ec4d7d8e29af3aaabac8611c85d97efca89f723a\""
#> [26] "  },"                                                              
#> [27] "  {"                                                               
#> [28] "    \"package\": \"nanonext\","                                    
#> [29] "    \"version_current\": \"1.5.1.9001\","                          
#> [30] "    \"hash_current\": \"e08cf94d86b23eed4a9968eee0042f4bcdd9cb8f\""
#> [31] "  },"                                                              
#> [32] "  {"                                                               
#> [33] "    \"package\": \"proffer\","                                     
#> [34] "    \"version_current\": \"0.2.2.9000\","                          
#> [35] "    \"hash_current\": \"161b9e8f8e528384deffa6b8dd6e7abeeb130531\""
#> [36] "  },"                                                              
#> [37] "  {"                                                               
#> [38] "    \"package\": \"withr\","                                       
#> [39] "    \"version_current\": \"3.0.2.9001\","                          
#> [40] "    \"hash_current\": \"59da5977ea1688e1c3064ab0c3eddb84c0be17d7\""
#> [41] "  }"                                                               
#> [42] "]"                                                                 
  # In subsequent snapshots, we have historical information about versions.
  record_versions(
    versions = versions,
    repo = repo
  )
  readLines(versions)
#>  [1] "["                                                                  
#>  [2] "  {"                                                                
#>  [3] "    \"package\": \"autometric\","                                   
#>  [4] "    \"version_current\": \"0.1.2.9000\","                           
#>  [5] "    \"hash_current\": \"13ec8d2cfba955081cf3e19eb5e74af07ba89bef\","
#>  [6] "    \"version_highest\": \"0.1.2.9000\","                           
#>  [7] "    \"hash_highest\": \"13ec8d2cfba955081cf3e19eb5e74af07ba89bef\"" 
#>  [8] "  },"                                                               
#>  [9] "  {"                                                                
#> [10] "    \"package\": \"crew\","                                         
#> [11] "    \"version_current\": \"1.0.0.9001\","                           
#> [12] "    \"hash_current\": \"7d460e26f4c45233bee15ff4db9564503ec4b21a\","
#> [13] "    \"version_highest\": \"1.0.0.9001\","                           
#> [14] "    \"hash_highest\": \"7d460e26f4c45233bee15ff4db9564503ec4b21a\"" 
#> [15] "  },"                                                               
#> [16] "  {"                                                                
#> [17] "    \"package\": \"crew.aws.batch\","                               
#> [18] "    \"version_current\": \"0.0.8\","                                
#> [19] "    \"hash_current\": \"136fedb9df30123b7c216ac75a66f29e8e94f2c1\","
#> [20] "    \"version_highest\": \"0.0.8\","                                
#> [21] "    \"hash_highest\": \"136fedb9df30123b7c216ac75a66f29e8e94f2c1\"" 
#> [22] "  },"                                                               
#> [23] "  {"                                                                
#> [24] "    \"package\": \"crew.cluster\","                                 
#> [25] "    \"version_current\": \"0.3.4.9000\","                           
#> [26] "    \"hash_current\": \"f9ecd38a1b8736157b06635f5e0cb43cb80c09a4\","
#> [27] "    \"version_highest\": \"0.3.4.9000\","                           
#> [28] "    \"hash_highest\": \"f9ecd38a1b8736157b06635f5e0cb43cb80c09a4\"" 
#> [29] "  },"                                                               
#> [30] "  {"                                                                
#> [31] "    \"package\": \"mirai\","                                        
#> [32] "    \"version_current\": \"2.1.0.9001\","                           
#> [33] "    \"hash_current\": \"ec4d7d8e29af3aaabac8611c85d97efca89f723a\","
#> [34] "    \"version_highest\": \"2.1.0.9001\","                           
#> [35] "    \"hash_highest\": \"ec4d7d8e29af3aaabac8611c85d97efca89f723a\"" 
#> [36] "  },"                                                               
#> [37] "  {"                                                                
#> [38] "    \"package\": \"nanonext\","                                     
#> [39] "    \"version_current\": \"1.5.1.9001\","                           
#> [40] "    \"hash_current\": \"e08cf94d86b23eed4a9968eee0042f4bcdd9cb8f\","
#> [41] "    \"version_highest\": \"1.5.1.9001\","                           
#> [42] "    \"hash_highest\": \"e08cf94d86b23eed4a9968eee0042f4bcdd9cb8f\"" 
#> [43] "  },"                                                               
#> [44] "  {"                                                                
#> [45] "    \"package\": \"proffer\","                                      
#> [46] "    \"version_current\": \"0.2.2.9000\","                           
#> [47] "    \"hash_current\": \"161b9e8f8e528384deffa6b8dd6e7abeeb130531\","
#> [48] "    \"version_highest\": \"0.2.2.9000\","                           
#> [49] "    \"hash_highest\": \"161b9e8f8e528384deffa6b8dd6e7abeeb130531\"" 
#> [50] "  },"                                                               
#> [51] "  {"                                                                
#> [52] "    \"package\": \"withr\","                                        
#> [53] "    \"version_current\": \"3.0.2.9001\","                           
#> [54] "    \"hash_current\": \"59da5977ea1688e1c3064ab0c3eddb84c0be17d7\","
#> [55] "    \"version_highest\": \"3.0.2.9001\","                           
#> [56] "    \"hash_highest\": \"59da5977ea1688e1c3064ab0c3eddb84c0be17d7\"" 
#> [57] "  }"                                                                
#> [58] "]"